Refactor code

+++ app/src/main/java/com/kyberswap/android/data/api/user/UserInfoEntity.kt
@@ -19,5 +19,5 @@ data class UserInfoEntity(
     @SerializedName("name")
     val name: String = "",
     @SerializedName("uid")
-    val uid: Int = 0
+    val uid: Long = 0
 )
